refactor(config): split built-in schema ownership
